CMS-1671: Abstract functions to files in middlewares/helpers
