+++ src/cms/src/middlewares/staff-portal-advisory-audit.js
@@ -5,6 +5,19 @@
  */
 
 const { queueAdvisoryEmail } = require("../helpers/taskQueue.js");
+const {
+  getNextAdvisoryNumber,
+  getNextRevisionNumber,
+} = require("./helpers/advisoryNumbers.js");
+const {
+  resolvePublishIntentStatus,
+  getAdvisoryStatusCode,
+} = require("./helpers/advisoryStatus.js");
+const {
+  archiveOldPublicAdvisoryAudit,
+  copyToPublicAdvisory,
+  isAdvisoryEqual,
+} = require("./helpers/advisoryData.js");
 
 module.exports = () => {
   /**
@@ -245,292 +258,3 @@ module.exports = () => {
     return await next();
   };
 };
-
-// HELPER FUNCTIONS
-
-async function getNextAdvisoryNumber() {
-  const results = await strapi
-    .documents("api::public-advisory-audit.public-advisory-audit")
-    .findMany({
-      sort: { advisoryNumber: "DESC" },
-      limit: 1,
-      fields: ["advisoryNumber"],
-    });
-  let maxAdvisoryNumber = results.length > 0 ? results[0].advisoryNumber : 0;
-  if (!maxAdvisoryNumber || maxAdvisoryNumber < 0) maxAdvisoryNumber = 0;
-  return ++maxAdvisoryNumber;
-}
-
-async function getNextRevisionNumber(advisoryNumber) {
-  const results = await strapi
-    .documents("api::public-advisory-audit.public-advisory-audit")
-    .findMany({
-      filters: {
-        advisoryNumber,
-      },
-      sort: { revisionNumber: "DESC" },
-      limit: 1,
-      fields: ["revisionNumber"],
-    });
-  let maxRevisionNumber = results.length > 0 ? results[0].revisionNumber : 0;
-  if (!maxRevisionNumber || maxRevisionNumber < 0) maxRevisionNumber = 0;
-  return ++maxRevisionNumber;
-}
-
-/**
- * Resolves publish intent into the advisoryStatus payload value to save in the DB.
- * Returns null when no status rewrite is needed.
- */
-async function resolvePublishIntentStatus(data) {
-  // Only resolve statuses when the request is trying to publish.
-  const requestedStatusCode = await getAdvisoryStatusCode(data?.advisoryStatus);
-  if (requestedStatusCode !== "PUB") {
-    return null;
-  }
-
-  // Apply date-based publishing rules: UNP (expired) > SCH (future) > PUB (active now).
-  const resolvedCode = resolvePublishStatusCode(data);
-  if (resolvedCode === requestedStatusCode) {
-    return null;
-  }
-
-  const resolvedStatus = await getAdvisoryStatusByCode(resolvedCode);
-  if (!resolvedStatus) {
-    return null;
-  }
-
-  return typeof data.advisoryStatus === "object" && data.advisoryStatus !== null
-    ? { id: resolvedStatus.id }
-    : resolvedStatus.documentId;
-}
-
-/**
- * Resolves publish target code based on expiry and posting dates.
- */
-function resolvePublishStatusCode(data) {
-  const now = new Date();
-  const expiryDate = data?.expiryDate ? new Date(data.expiryDate) : null;
-  const advisoryDate = data?.advisoryDate ? new Date(data.advisoryDate) : null;
-
-  // If the expiry date is now or in the past, use "Unpublished" status
-  if (expiryDate && !Number.isNaN(expiryDate.getTime()) && expiryDate <= now) {
-    return "UNP";
-  }
-
-  // If the posting date is in the future, use "Scheduled" status
-  if (
-    advisoryDate &&
-    !Number.isNaN(advisoryDate.getTime()) &&
-    advisoryDate > now
-  ) {
-    return "SCH";
-  }
-
-  // Default: advisory is active now, so use "Published" status
-  return "PUB";
-}
-
-/**
- * Reads advisory status code from supported relation payload shapes.
- */
-async function getAdvisoryStatusCode(statusValue) {
-  // Accept relation values in different shapes (object with code/id/documentId, or documentId string).
-  if (!statusValue) return null;
-
-  if (typeof statusValue === "object" && statusValue !== null) {
-    // If statusValue already has the code, return it directly
-    if (statusValue.code) {
-      return statusValue.code;
-    }
-
-    // Look up the code based on the provided id or documentId
-    if (statusValue.documentId) {
-      const status = await strapi
-        .documents("api::advisory-status.advisory-status")
-        .findOne({
-          documentId: statusValue.documentId,
-          fields: ["code"],
-        });
-      return status?.code ?? null;
-    }
-
-    if (statusValue.id) {
-      const status = await strapi
-        .documents("api::advisory-status.advisory-status")
-        .findFirst({
-          filters: { id: statusValue.id },
-          fields: ["code"],
-        });
-      return status?.code ?? null;
-    }
-  }
-
-  // If it's a string, treat it as a documentId and look up the code
-  if (typeof statusValue === "string") {
-    const status = await strapi
-      .documents("api::advisory-status.advisory-status")
-      .findOne({
-        documentId: statusValue,
-        fields: ["code"],
-      });
-    return status?.code ?? null;
-  }
-
-  return null;
-}
-
-/**
- * Looks up advisory-status metadata by code.
- */
-async function getAdvisoryStatusByCode(code) {
-  // Resolve status metadata once we know the target code.
-  return await strapi
-    .documents("api::advisory-status.advisory-status")
-    .findFirst({
-      filters: { code },
-      fields: ["id", "documentId", "code"],
-    });
-}
-
-/**
- * Creates an archived copy of the provided advisory data.
- */
-async function archiveOldPublicAdvisoryAudit(data) {
-  delete data.id;
-  delete data.documentId;
-  delete data.updatedBy;
-  delete data.createdBy;
-  delete data.links; // keep links connected to the latest revision, not the archived version
-  data.publishedAt = null;
-  data.isLatestRevision = false;
-
-  try {
-    await strapi
-      .documents("api::public-advisory-audit.public-advisory-audit")
-      .create({ data: data });
-  } catch (error) {
-    strapi.log.error(
-      `error creating public-advisory-audit ${data.advisoryNumber}...`,
-      error,
-    );
-  }
-}
-
-async function savePublicAdvisory(publicAdvisory) {
-  delete publicAdvisory.updatedBy;
-  delete publicAdvisory.createdBy;
-  const isExist = await strapi
-    .documents("api::public-advisory.public-advisory")
-    .findFirst({
-      filters: {
-        advisoryNumber: publicAdvisory.advisoryNumber,
-      },
-      fields: ["advisoryNumber"],
-    });
-  if (publicAdvisory.advisoryStatus.code === "PUB") {
-    publicAdvisory.publishedAt = new Date();
-    if (isExist) {
-      try {
-        publicAdvisory.id = isExist.id;
-        publicAdvisory.documentId = isExist.documentId;
-        strapi.log.info(
-          `updating public-advisory advisoryNumber ${publicAdvisory.advisoryNumber}...`,
-        );
-        await strapi.documents("api::public-advisory.public-advisory").update({
-          documentId: isExist.documentId,
-          data: publicAdvisory,
-        });
-      } catch (error) {
-        strapi.log.error(
-          `error updating public-advisory advisoryNumber ${publicAdvisory.advisoryNumber}...`,
-          error,
-        );
-      }
-    } else {
-      try {
-        delete publicAdvisory.id;
-        delete publicAdvisory.documentId;
-        strapi.log.info(
-          `creating public-advisory advisoryNumber ${publicAdvisory.advisoryNumber}...`,
-        );
-        await strapi.documents("api::public-advisory.public-advisory").create({
-          data: publicAdvisory,
-        });
-      } catch (error) {
-        strapi.log.error(
-          `error creating public-advisory ${publicAdvisory.id}...`,
-          error,
-        );
-      }
-    }
-  } else if (isExist) {
-    try {
-      strapi.log.info(
-        `deleting public-advisory advisoryNumber ${publicAdvisory.advisoryNumber}...`,
-      );
-      await strapi
-        .documents("api::public-advisory.public-advisory")
-        .delete({ documentId: isExist.documentId });
-    } catch (error) {
-      strapi.log.error(
-        `error deleting public-advisory ${publicAdvisory.id}...`,
-        error,
-      );
-    }
-  }
-}
-
-async function copyToPublicAdvisory(newPublicAdvisory) {
-  if (newPublicAdvisory.isLatestRevision && newPublicAdvisory.advisoryStatus) {
-    const triggerStatuses = ["PUB", "UNP"];
-    if (triggerStatuses.includes(newPublicAdvisory.advisoryStatus.code)) {
-      await savePublicAdvisory(newPublicAdvisory);
-    }
-  }
-}
-
-function isAdvisoryEqual(newData, oldData) {
-  const fieldsToCompare = {
-    title: null,
-    description: null,
-    isSafetyRelated: null,
-    listingRank: null,
-    note: null,
-    advisoryDate: null,
-    effectiveDate: null,
-    endDate: null,
-    expiryDate: null,
-    accessStatus: {},
-    eventType: {},
-    urgency: {},
-    standardMessages: [],
-    protectedAreas: [],
-    advisoryStatus: {},
-    links: [],
-    regions: [],
-    sections: [],
-    managementAreas: [],
-    sites: [],
-    fireCentres: [],
-    fireZones: [],
-    naturalResourceDistricts: [],
-    isAdvisoryDateDisplayed: null,
-    isEffectiveDateDisplayed: null,
-    isEndDateDisplayed: null,
-    isUpdatedDateDisplayed: null,
-    modifiedBy: null,
-  };
-
-  for (const key of Object.keys(fieldsToCompare)) {
-    if (Array.isArray(oldData[key])) {
-      oldData[key] = oldData[key].map((x) => x.id).sort();
-      if (newData[key]) newData[key].sort();
-    } else {
-      if (typeof oldData[key] === "object" && oldData[key])
-        oldData[key] = oldData[key].id;
-    }
-    if (JSON.stringify(newData[key]) != JSON.stringify(oldData[key]))
-      return false;
-  }
-  return true;
-}
